FileCache extracted to spm package
