+++ FileCache/Sources/FileCache/FileCache.swift
@@ -1,9 +1,5 @@
-//
-//  FileCache.swift
-//  YaToDo
-//
-//  Created by Danil Masnaviev on 17/06/24.
-//
+// The Swift Programming Language
+// https://docs.swift.org/swift-book
 
 import Foundation
 
@@ -14,26 +10,24 @@ enum FileCacheError: Error {
     case parseFailed
 }
 
-final class FileCache {
-    private(set) var todos: [TodoItem] = []
+public final class FileCache<T: Cacheable> {
+    private(set) var items: [T] = []
 
-    func addTodo(_ todo: TodoItem) {
-        if let index = todos.firstIndex(where: { $0.id == todo.id }) {
-            todos[index] = todo
+    func addItem(_ item: T) {
+        if let index = items.firstIndex(where: { $0.id == item.id }) {
+            items[index] = item
         } else {
-            todos.append(todo)
+            items.append(item)
         }
     }
 
     func deleteTodo(_ id: String) {
-        todos.removeAll { $0.id == id }
+        items.removeAll { $0.id == id }
     }
-}
 
-extension FileCache {
     func saveJSON(filename: String) throws {
         let fileURL = try getFileURL(filename)
-        let jsonItems = todos.map({ $0.json })
+        let jsonItems = items.map({ $0.json })
 
         do {
             let data = try JSONSerialization.data(withJSONObject: jsonItems, options: [])
@@ -50,8 +44,8 @@ extension FileCache {
             let data = try Data(contentsOf: fileURL)
             if let jsonItems = try JSONSerialization.jsonObject(with: data, options: []) as? [[String: Any]] {
                 for json in jsonItems {
-                    if let todoItem = TodoItem.parse(json: json) {
-                        addTodo(todoItem)
+                    if let item = T.parse(json: json) {
+                        addItem(item)
                     } else {
                         throw FileCacheError.parseFailed
                     }
@@ -64,7 +58,7 @@ extension FileCache {
 
     func saveCSV(filename: String) throws {
         let fileURL = try getFileURL(filename)
-        let csvItems = todos.map({ $0.csv }).joined(separator: "\n")
+        let csvItems = items.map({ $0.csv }).joined(separator: "\n")
 
         do {
             try csvItems.write(to: fileURL, atomically: true, encoding: .utf8)
@@ -81,8 +75,8 @@ extension FileCache {
             let csvLines = data.components(separatedBy: "\n")
 
             for line in csvLines {
-                if let todoItem = TodoItem.parse(csv: line) {
-                    addTodo(todoItem)
+                if let item = T.parse(csv: line) {
+                    addItem(item)
                 }
             }
         } catch {
