+++ YaToDo.xcodeproj/xcuserdata/danilmasnaviev.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,7 +7,20 @@
 		<key>YaToDo.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>7</integer>
+			<integer>3</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>3C9206192C25ECC0002A8272</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>3CF661992C2085CC00BD719F</key>
+		<dict>
+			<key>primary</key>
+			<true/>
 		</dict>
 	</dict>
 </dict>
